==> notebooks/04_baseline_with_example.ipynb <==
import sys
from pathlib import Path

# Add project root to Python path
project_root = Path.cwd().parent
if str(project_root) not in sys.path:
    sys.path.insert(0, str(project_root))

import dspy
from src.modules_v2 import IntegratedText2SQLModule, ViewSelectorModule, EntityResolverModule
from src.data_loader import load_golden_dataset, load_snowflake_views, create_dspy_examples
from src.metrics import calculate_metrics, print_evaluation_report


lm = dspy.LM("azure/gpt-4o")
dspy.configure(lm=lm)
print("✅ DSPy configured")

golden_df = load_golden_dataset("../data/golden_dataset_v1.csv")
snowflake_views = load_snowflake_views("../data/snowflake_view.json")
examples = create_dspy_examples(golden_df)

examples[1]

entity_resolve_module = EntityResolverModule(config={})
view_selector_module = ViewSelectorModule(config = {}, candidate_views = snowflake_views)

entity_resolve_prediction = entity_resolve_module(
                question=examples[1]['question'] 
            )

entity_resolve_prediction

view_selector_prediction = view_selector_module(
                question=examples[1]['question'],
                resolved_entities=entity_resolve_prediction.resolved_entities,
                conversation_history=examples[1]['conversation_history']
            )

print(view_selector_prediction.selected_views)
print(examples[1]['expected_views'])
